Extreme Refactor (add build step)
- Add build step
- Fix all the problems that are associated with adding a build step.... mostly tests
- Restructure some of testing
- Upgrade to AIK 0.8.0
- Upgrade all dep
