+++ package/package.json
@@ -2,11 +2,7 @@
 	"name": "astro-theme-provider",
 	"version": "0.1.1",
 	"description": "Easily create theme integrations for Astro",
-	"keywords": [
-		"astro",
-		"withastro",
-		"astro-integration"
-	],
+	"keywords": ["astro", "withastro", "astro-integration"],
 	"author": "Bryce Russell",
 	"license": "Unlicense",
 	"repository": {
@@ -17,6 +13,7 @@
 	"bugs": "https://github.com/brycerussell/astro-theme-provider/issues",
 	"homepage": "https://astro-theme-provider.netlify.app/",
 	"scripts": {
+		"test": "node --test",
 		"dev": "tsup --watch",
 		"build": "tsup"
 	},
@@ -27,9 +24,7 @@
 			"default": "./dist/index.js"
 		}
 	},
-	"files": [
-		"dist"
-	],
+	"files": ["dist"],
 	"peerDependencies": {
 		"@astrojs/db": ">=0.8.0",
 		"astro": ">=3"
@@ -40,13 +35,13 @@
 		}
 	},
 	"devDependencies": {
-		"@astrojs/db": "^0.8.7",
-		"@types/node": "^20.11.29",
+		"@astrojs/db": "^0.9.5",
+		"@types/node": "^20.11.30",
 		"tsup": "^8.0.2",
 		"typescript": "^5.4.3"
 	},
 	"dependencies": {
-		"astro-integration-kit": "^0.6.0",
+		"astro-integration-kit": "^0.8.0",
 		"astro-pages": "^0.2.1",
 		"astro-public": "^0.0.4",
 		"callsites": "^4.1.0",
