+++ package/src/internal/error-map.ts
@@ -47,7 +47,7 @@ export const errorMap: ZodErrorMap = (baseError, ctx) => {
 						.map(([key, error]) =>
 							key === baseErrorPath
 								? // Avoid printing the key again if it's a base error
-								  `> ${getTypeOrLiteralMsg(error)}`
+									`> ${getTypeOrLiteralMsg(error)}`
 								: `> ${prefix(key, getTypeOrLiteralMsg(error))}`,
 						),
 				)
