+++ package/tests/mock/package/src/assets/ai.png
