+++ package/tests/mock/package/src/assets/cursed.png
