+++ package/tests/mock/package/src/assets/starlight.png
