+++ package/tests/mock/package/src/pages/api/json.json.ts
