+++ package/tests/mock/package/src/pages/cats/[...cat].astro
