+++ package/tests/mock/package/src/pages/cats/index.astro
