+++ package/tests/mock/package/src/pages/index.astro
