+++ package/tests/mock/project/env.d.ts
@@ -1,0 +1,3 @@
+/// <reference path="../.astro/types.d.ts" />
+/// <reference types="astro/client" />
+/// <reference types="../../.astro/my-theme.d.ts" />
