+++ package/tests/mock/project/src/env.d.ts
