+++ package/tests/unit/provider.test.js
@@ -2,11 +2,12 @@ import * as assert from "node:assert/strict";
 import { dirname, resolve } from "node:path";
 import { afterEach, describe, it, mock } from "node:test";
 import { fileURLToPath, pathToFileURL } from "node:url";
-import { addDts, addIntegration, addVirtualImports, watchIntegration } from "astro-integration-kit/utilities";
-import _defineTheme from "../index.js";
+// Mock imports
+import { addDts, addIntegration, addVirtualImports, watchIntegration } from "astro-integration-kit";
+import _defineTheme from "../../dist/index.js";
 
 const thisFile = fileURLToPath(import.meta.url).toString();
-const playgroundDir = resolve(dirname(thisFile), "mock");
+const playgroundDir = resolve(dirname(thisFile), "../mock");
 const packageRoot = resolve(playgroundDir, "package");
 const packageEntrypoint = resolve(packageRoot, "index.ts");
 const projectRoot = resolve(playgroundDir, "project");
@@ -46,6 +47,7 @@ describe("defineTheme", () => {
 	mock.fn(addDts, () => {});
 	mock.fn(addIntegration, () => {});
 	mock.fn(addVirtualImports, () => {});
+	mock.fn(watchIntegration, () => {});
 	mock.fn(watchIntegration, () => {});
 
 	afterEach(() => {
