+++ package/tests/unit/virtual.test.js
@@ -1,7 +1,7 @@
 import * as assert from "node:assert/strict";
 import { isAbsolute } from "node:path";
 import { describe, it } from "node:test";
-import { resolveModuleObject } from "../../src/utils/virtual.js";
+import { resolveModuleObject } from "../../dist/utils/virtual.js";
 
 const moduleObject = {
 	imports: ["./styles.css"],
